Create serverNegotiations
I passed false into the play_Nim function which is what I think Scott is working on. I think that means that it is client's turn first. There are probably a lot of errors in the code. If you find any feel free to fix them or tell me to.